requirements.txt
Flask
gunicorn
boto3
PyYAML
Markdown
